Improve handling of globals of class types
Since these values are of Why3 type `reference', we'll need to know they
belong to a particular class type in order to use them (access or update
fields).  This commit adds (to each method and loop body) preconditions
and loop invariants of the form `hasClassTypeK s s.global_var' for each
global of class type K.

In the future, these conditions should be added only if the global
variable occurs in a method body.