Add `diverges' to methods that may diverge
We are only concerned with partial correctness and usually flag functions in
Why3 that contain loops (or call other functions that may diverge) with a
`diverges' clause to avoid VCs related to termination.

Previously, this `diverges' annotation had to be added manually---annoying and
error-prone, especially when done using scripts.  This commit addresses the
issue: WhyRel will now add a diverges annotation to specs in modules and
bimodules automatically.

However, the pretty-printer in Why3.1.3.3 will not print out the `diverges'
annotation correctly (it prints "@ diverge" instead of "diverges").  The
util/post-process.sed file has been updated to perform the appropriate
substitution.