Handle image expressions for arrays of objects and regions
WhyRel axiomatizes image expressions---for each field f, we generate a Why3
function img_f of type state -> rgn -> rgn and an axiom that defines it.  This
commit handles the case when f is the slots field of an array-like class and
has type (K array) for some class K or (rgn array).  Previously, img_f here
was axiomatized as being the empty region.