Fix bug in computing effect of a method call
When computing the effect of the command [x := m(args)], we substitute
args for the formal parameters of m in m's frame (works because we
restrict each argument to be a variable).  However, m's effect clause
could include rw result or rw result`any (and so on), but we weren't
substituting x for result.  This commit fixes the issue.