Use setters for fields and globals
Encoding changed to use setters.  The field update `o.f := g` used to get
translated to (assuming g' is what g translates to):

  s.heap.f <- M.add o g' s.heap.f

Now, it gets translated to:

  set_f s o g'

There are a few bugs to fix.  The function fields_written_bimethod no longer
computes the actual fields written by a bimethod.  It operates on the Why3
parse trees we generate, which have changed now.  Both fields_written and
fields_written_bimethod *should* be operating at the WhyRel level and not the
Why3 level.