Add: "havoc on the right such that"
Biprograms can now include the statement "havoc x { r }", where x is a
variable and r is a relation formula.  This will havoc x on the right and then
assume r.  Currently, "havoc x { r }" is treated as syntactic sugar for

  assert { exists x. r }; (skip | havoc x); assume { r }

A possible generalization is the statement "havoc x|x' { r }" which is
syntactic sugar for

  assert { forall x. exists x'. r }; (havoc x | havoc x'); assume { r }

The right projection of havoc x { r } is simply havoc x.  Its left projection
is just skip.